// api/src/domain/interfaces/common/request.ts
interface IRequestBody {
  nomeCliente: string
  enderecoCliente: string
  bairroCliente: string
  cidadeCliente: string
  municipioCliente: string
  estadoCliente: string
  idProposta: 431231
  metodo: string
  comportamento: string
  visualizacaoHumana: string
  custo: string
  condicoesPagamento: string
  cargoCliente: string
  nomeUsuario: string
  dataCriacaoDocumento: string
  ultimaDataAlteracaoDocumento: string
}

export default IRequestBody